refactor(server): extract command client state
