+++ server/proliferate/server/cloud/worker/control/checks.py
@@ -12,7 +12,7 @@ from proliferate.db.store.cloud_sync import commands as commands_store
 from proliferate.db.store.cloud_sync import targets as targets_store
 from proliferate.db.store.cloud_sync import worker_control as worker_control_store
 from proliferate.db.store.cloud_sync import worker_exposures as worker_exposures_store
-from proliferate.server.cloud.commands import service as command_service
+from proliferate.server.cloud.commands.client_state import expire_stale_client_commands_for_target
 from proliferate.server.cloud.errors import CloudApiError
 from proliferate.server.cloud.event_logging import log_cloud_event
 from proliferate.server.cloud.live.service import publish_command_status_after_commit
@@ -63,7 +63,7 @@ async def check_worker_control(
         _require_current_worker_target(target)
 
         now = utcnow()
-        expired_commands = await command_service.expire_stale_client_commands_for_target(
+        expired_commands = await expire_stale_client_commands_for_target(
             db,
             target_id=auth.target_id,
         )
